CherryBlossom-common/src/main/java/com/cherryblossom/baomidou/model/Phone.java
package com.cherryblossom.baomidou.model;

import com.baomidou.mybatisplus.activerecord.Model;
import com.baomidou.mybatisplus.annotations.TableName;

import java.io.Serializable;

/**
 * Created by nieqiurong on 2016/12/12.
 */
@TableName(value = "t_phone")
public class Phone extends Model<Phone> {
    private Long id;
    private String userName;
    private Integer age;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getUserName() {
        return userName;
    }

    public void setUserName(String userName) {
        this.userName = userName;
    }

    public Integer getAge() {
        return age;
    }

    public void setAge(Integer age) {
        this.age = age;
    }

    @Override
    protected Serializable pkVal() {
        return id;
    }

    public Phone() {
    }

    public Phone(String userName, Integer age) {
        this.userName = userName;
        this.age = age;
    }

    public Phone(Long id, String userName, Integer age) {
        this.id = id;
        this.userName = userName;
        this.age = age;
    }

    @Override
    public String toString() {
        return "Phone{" +
                "id=" + id +
                ", userName='" + userName + '\'' +
                ", age=" + age +
                '}';
    }
}
